LLM_Security_Gateway/requirements.txt
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
spacy==3.7.2
pyyaml==6.0.1
pandas==2.0.3
tabulate==0.9.0
ollama==0.1.6
requests==2.31.0
matplotlib==3.7.2
seaborn==0.12.2